SNAPDEALTECH-81738 removing unwanted files

+++ HealthCheckApp/src/main/java/com/snapdeal/healthcheck/app/model/DownTimeUIData.java
@@ -3,25 +3,25 @@ package com.snapdeal.healthcheck.app.model;
 public class DownTimeUIData {
 	
 	String id;
-	int leftMargin;
-	int width;
+	double leftMargin;
+	double width;
 	String downTime;
 	String upTime;
 	String totalTime;
 
-	public int getLeftMargin() {
+	public double getLeftMargin() {
 		return leftMargin;
 	}
 
-	public void setLeftMargin(int leftMargin) {
+	public void setLeftMargin(double leftMargin) {
 		this.leftMargin = leftMargin;
 	}
 
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
 
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
